requirements.txt
opencv-python
numpy
ultralytics
cvzone
matplotlib
Pillow
pyinstaller
pandas